ai_scripts/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
attrs==24.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
cymem==2.0.8
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
importlib_metadata==8.5.0
Jinja2==3.1.4
jiter==0.5.0
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
murmurhash==1.0.10
numpy==1.26.4
openai==1.44.1
packaging==24.1
preshed==3.0.9
pydantic==2.9.1
pydantic_core==2.23.3
pyee==11.1.1
Pygments==2.18.0
pyppeteer==2.0.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.8.1
setuptools==74.1.2
shellingham==1.5.4
smart-open==7.0.4
sniffio==1.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.5
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
urllib3==1.26.20
wasabi==1.1.3
weasel==0.4.1
websockets==10.4
wrapt==1.16.0
yarl==1.11.1
zipp==3.20.1
